# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standard error response schema published in OpenAPI.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LimitsResponse": {
        "properties": {
          "max_upload_size_bytes": {
            "title": "Max Upload Size Bytes",
            "type": "integer"
          },
          "max_upload_size_mb": {
            "title": "Max Upload Size Mb",
            "type": "integer"
          }
        },
        "required": [
          "max_upload_size_mb",
          "max_upload_size_bytes"
        ],
        "title": "LimitsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Agent Knowledge Repository",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/config/limits": {
      "get": {
        "description": "Client-facing upload limits from the live configuration.",
        "operationId": "config_limits_api_v1_config_limits_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Config Limits"
      }
    },
...
  }
}